pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "BookStore API"
authors = ["Fernando"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = ">=5.2,<6.0"
djangorestframework = ">=3.16.0,<4.0.0"
django-rest-framework = ">=0.1.0,<0.2.0"
django-extensions = ">=4.1,<5.0"
django-debug-toolbar = ">=5.2.0,<6.0.0"
psycopg2-binary = "^2.9.0"
gitpython = "^3.1.44"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
factory-boy = "^3.3.3"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bookstore.settings"
python_files = ["test_*.py", "*_tests.py"]